requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
loguru==0.7.2
pillow==10.1.0
openai>=1.0.0